Allow specifying variable font range descriptors in either direction
Style, stretch and weight descriptor allow specifying ranges for the
variable font that the src: descriptor is referencing. The range
definitions allow both the smaller or larger value to go first and need
to be swapped by the UA.

Keep range value lists in reversed order in the computed style, but swap
them in the FontFace implementation to become meaningful values.This
keeps serialisation intact as specified in the @font - face declaration(
as opposed to reversing the list at the CSS parsing level) and matches
Firefox' behaviour.

Fixed: 1063867
Change-Id: I5ae7bc5f3e3a75fa9dce09ca0a66dd083ef29b0b